// src/integrations/contentful/types/product-fields.type.ts
export type ContentfulProductFields = {
  sku?: string;
  name?: string;
  brand?: string;
  model?: string;
  category?: string;
  color?: string;
  price?: number;
  currency?: string;
  stock?: number;
};

export type ContentfulEntrySys = {
  id: string;
  createdAt: string;
  updatedAt: string;
};